+++ core/src/main/java/com/sap/ai/sdk/core/model/AiExecution.java
@@ -18,6 +18,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonValue;
+import com.google.common.annotations.Beta;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
@@ -30,7 +31,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 /** Execution that may generate artifacts */
-// CHECKSTYLE:OFF
+@Beta // CHECKSTYLE:OFF
 public class AiExecution
 // CHECKSTYLE:ON
 {