+++ core/src/main/java/com/sap/ai/sdk/core/model/BckndResourceGetResponse.java
@@ -16,6 +16,7 @@ import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.google.common.annotations.Beta;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -26,7 +27,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 /** BckndResourceGetResponse */
-// CHECKSTYLE:OFF
+@Beta // CHECKSTYLE:OFF
 public class BckndResourceGetResponse
 // CHECKSTYLE:ON
 {