+++ core/src/main/java/com/sap/ai/sdk/core/model/RTADeploymentDetails.java
@@ -16,6 +16,7 @@ import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.google.common.annotations.Beta;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.NoSuchElementException;
@@ -25,7 +26,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 /** RTADeploymentDetails */
-// CHECKSTYLE:OFF
+@Beta // CHECKSTYLE:OFF
 public class RTADeploymentDetails
 // CHECKSTYLE:ON
 {