+++ docs/release-notes/release_notes.md
@@ -10,6 +10,7 @@
 
 - `ChatMessage`, as well as new `MultiChatMessage`, are now subtypes of new interface `ChatMessagesInner`.
   Most variables or methods previously typed as `ChatMessage` in `model` package are now typed as `ChatMessagesInner`. 
+- Add missing `@Beta` annotations to all `com.sap.ai.sdk.core.client` and `com.sap.ai.sdk.core.model` classes.
 
 ### ✨ New Functionality